sampler: remove testing.go and move into coresampler_test.go
